CARDS-2332: Add automatic saving and checkin of forms after a period of inactivity Code refactoring
